Got multiple tips to load.
Currently 5 tips load and no more after that. Need to figure out how to maintain a certain number of tips (i.e., if the number of tips falls below a threshold, then load more (which just requires reloading the webView). It's probably easiest to accomplish this by just telling the webView to reload every time a tip is liked or disliked; however, then we'd still need to get an initial number of tips, and these logics might happen in separate places which is less than ideal though not terrible. Alternatively, we might be able to add a listener to the number of tips in the stack so that whenever the number is below a threshold, another tip is loaded. This would be more complicated but have the benefit of keeping the code that maintains the number of tips in one place. In either case, we might need to pass the webView to whatever class is executing the maintenance logic so it can appropriately reload the page.

Also currently when the tips are being loaded, each new tip goes on the top of the stack, which is not good. We need to figure out how to make sure it goes to the bottom so that users aren't interrupted when a new one loads.