New tips now load behind older tips. Added logic to remove surrounding quotes from tips.
When new tips are loaded, previously voted on tips reappear as a result of the adapter being re-set; however, not re-setting the adapter causes the tips to not be interactable. Options: figure out how to not have to re-set the adapter while maintaining interactability; waiting some time after a tip is voted on might work; rewrite most of the CardContainer class. The last option seems least desirable. Next time I'll look into the middle and continue doing some research on the first.